restructure and add dt docs
