+++ docs/densitytensor.rst
@@ -1,0 +1,14 @@
+densitytensor
+=======================
+
+.. toctree::
+
+    densitytensor/kraus
+    densitytensor/get_params_to_densitytensor_func
+    densitytensor/partial_trace
+    densitytensor/get_densitytensor_to_expectation_func
+    densitytensor/get_densitytensor_to_sampled_expectation_func
+    densitytensor/densitytensor_to_measurement_probabilities
+    densitytensor/densitytensor_to_measured_densitytensor
+    densitytensor/statetensor_to_densitytensor
+
