+++ docs/densitytensor/densitytensor_to_measured_densitytensor.rst
@@ -1,0 +1,5 @@
+densitytensor_to_measured_densitytensor
+==============================================
+
+.. autofunction:: qujax.densitytensor_to_measured_densitytensor
+
