+++ docs/densitytensor/densitytensor_to_measurement_probabilities.rst
@@ -1,0 +1,5 @@
+densitytensor_to_measurement_probabilities
+==============================================
+
+.. autofunction:: qujax.densitytensor_to_measurement_probabilities
+
