+++ docs/densitytensor/get_densitytensor_to_sampled_expectation_func.rst
@@ -1,0 +1,5 @@
+get_densitytensor_to_sampled_expectation_func
+================================================
+
+.. autofunction:: qujax.get_densitytensor_to_sampled_expectation_func
+
