+++ docs/densitytensor/get_params_to_densitytensor_func.rst
@@ -1,0 +1,5 @@
+get_params_to_densitytensor_func
+===================================
+
+.. autofunction:: qujax.get_params_to_densitytensor_func
+
