+++ docs/densitytensor/kraus.rst
@@ -1,0 +1,5 @@
+kraus
+=======================
+
+.. autofunction:: qujax.kraus
+
