+++ docs/densitytensor/partial_trace.rst
@@ -1,0 +1,5 @@
+partial_trace
+===============================
+
+.. autofunction:: qujax.partial_trace
+
