+++ docs/densitytensor/statetensor_to_densitytensor.rst
@@ -1,0 +1,5 @@
+statetensor_to_densitytensor
+===============================
+
+.. autofunction:: qujax.statetensor_to_densitytensor
+
