+++ qujax/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.2.9'
+__version__ = '0.3.0'
